# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CCE API",
    "description": "Code context engine: indexing, search, relations and tool endpoints",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/project/{id}/config": {
      "put": {
        "tags": [
          "Project"
        ],
        "summary": "Handle update project config request",
        "operationId": "handle_update_project_config",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectConfigUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{id}/dead-letter/retry": {
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "Handle a manual dead-letter truncate-retry request",
        "operationId": "handle_dead_letter_retry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterRetryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{id}/index": {
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "Handle project indexing request",
        "operationId": "handle_project_index",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectIndexResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{id}/reload": {
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "Handle reload project config request (hot reload)",
        "operationId": "handle_reload_project_config",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectConfigReloadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeadLetterRetryResponse": {
        "type": "object",
        "description": "Dead-letter truncate-retry response",
        "required": [
          "success",
          "retried",
          "succeeded",
          "still_failed",
          "truncated_chunks",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "retried": {
            "type": "integer",
            "description": "Files re-chunked from disk and re-stored.",
            "minimum": 0
          },
          "still_failed": {
            "type": "integer",
            "description": "Files retried that failed again; they stay in the dead letter state.",
            "minimum": 0
          },
          "succeeded": {
            "type": "integer",
            "description": "Files whose Embedding module succeeded after the truncate retry.",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "truncated_chunks": {
            "type": "integer",
            "description": "Chunks truncated during the retry.",
            "minimum": 0
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "description": "Error detail structure",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Common error response",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ProjectConfigReloadResponse": {
        "type": "object",
        "description": "Response for reloading project config from file",
        "required": [
          "success",
          "message",
          "project_id",
          "config_version"
        ],
        "properties": {
          "config_version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "message": {
            "type": "string"
          },
          "project_id": {
            "type": "integer",
            "format": "int64"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "ProjectConfigUpdateRequest": {
        "type": "object",
        "description": "Request body for updating project config",
        "required": [
          "config"
        ],
        "properties": {
          "config": {
            "type": "object",
            "description": "Project-level configuration (partial config, free-form)"
          }
        }
      },
      "ProjectConfigUpdateResponse": {
        "type": "object",
        "description": "Response for updating project config",
        "required": [
          "success",
          "hot_reload_applied",
          "message"
        ],
        "properties": {
          "hot_reload_applied": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ProjectIndexResponse": {
        "type": "object",
        "description": "Project index trigger response",
        "required": [
          "success",
          "project_id",
          "project_name",
          "indexed_files",
          "total_entities",
          "total_vectors",
          "elapsed_ms"
        ],
        "properties": {
          "elapsed_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "indexed_files": {
            "type": "integer",
            "minimum": 0
          },
          "project_id": {
            "type": "integer",
            "format": "int64"
          },
          "project_name": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "total_entities": {
            "type": "integer",
            "minimum": 0
          },
          "total_vectors": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}